requirements.txt
# Web Framework
Flask==2.3.2
Flask-CORS==4.0.0
gunicorn==21.2.0

# Google Earth Engine and Authentication
earthengine-api==0.1.363
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-cloud-storage>=2.0.0

# Deep Learning
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.0.74

# Geospatial Libraries  
rasterio==1.3.7
geopandas==0.13.2
shapely==2.0.1
fiona==1.9.4
pyproj==3.6.0
geojson==3.0.1

# Scientific Computing
numpy==1.24.3
scipy==1.11.1
Pillow==10.0.0
scikit-image==0.21.0

# Data Processing and Utilities
python-dotenv==1.0.0
requests==2.31.0
pandas>=1.3.0 